First stab at an example for Buck
As things stand, these examples can be run from an IDE with the CWD set
to the project root. They provide hot-deploy for the example1 module,
which has been lightly modified with a BUCK file, and a maven step to
dump dependencies to a lib folder that Buck can reference.

This shouldn't be coupled with vertx-hot-example1. In doing so, it
mixes Buck and Maven examples, and doesn't provide a realistic Buck
environment. I'll build out a fully-fledged structure in example-buck
as I get time.